+++ api/src/org/labkey/api/query/QueryRowReference.java
@@ -21,12 +21,12 @@ public class QueryRowReference
     final @NotNull String _queryName;
     final @NotNull List<Pair<FieldKey, Object>> _pkFilters;
 
-    public QueryRowReference(@NotNull Container c, @NotNull SchemaKey schemaKey, @NotNull String queryName, @NotNull Enum pkCol, int pkValue)
+    public QueryRowReference(@NotNull Container c, @NotNull SchemaKey schemaKey, @NotNull String queryName, @NotNull Enum pkCol, long pkValue)
     {
         this(c, schemaKey, queryName, List.of(Pair.of(FieldKey.fromParts(pkCol), pkValue)));
     }
 
-    public QueryRowReference(@NotNull Container c, @NotNull SchemaKey schemaKey, @NotNull String queryName, @NotNull FieldKey pkCol, int pkValue)
+    public QueryRowReference(@NotNull Container c, @NotNull SchemaKey schemaKey, @NotNull String queryName, @NotNull FieldKey pkCol, long pkValue)
     {
         this(c, schemaKey, queryName, List.of(Pair.of(pkCol, pkValue)));
     }
