+++ assay/api-src/org/labkey/api/assay/dilution/DilutionManager.java
@@ -139,7 +139,7 @@ public class DilutionManager
         return (Integer)newFields.get("RowId");
     }
 
-    public static List<DilutionDataRow> getDilutionDataRows(int runId, int plateNumber, String name, Container container, boolean filterReplicate)
+    public static List<DilutionDataRow> getDilutionDataRows(long runId, long plateNumber, String name, Container container, boolean filterReplicate)
     {
         SimpleFilter filter = SimpleFilter.createContainerFilter(container);
         filter.addCondition(FieldKey.fromString("runId"), runId);
@@ -204,7 +204,7 @@ public class DilutionManager
     }
 
     @Nullable
-    public NabSpecimen getNabSpecimen(int rowId)
+    public NabSpecimen getNabSpecimen(long rowId)
     {
         Filter filter = new SimpleFilter(FieldKey.fromString("RowId"), rowId);
         return getNabSpecimen(filter);
@@ -229,7 +229,7 @@ public class DilutionManager
         return null;
     }
 
-    public List<NabSpecimen> getNabSpecimens(List<Integer> rowIds)
+    public List<NabSpecimen> getNabSpecimens(List<Long> rowIds)
     {
         TableInfo tableInfo = getSchema().getTable(NAB_SPECIMEN_TABLE_NAME);
         Filter filter = new SimpleFilter(new SimpleFilter.InClause(FieldKey.fromString("RowId"), rowIds));
@@ -278,8 +278,8 @@ public class DilutionManager
         try (DbScope.Transaction transaction = scope.ensureTransaction())
         {
             // Get dataIds that match the ObjectUri and make filter on NabSpecimen
-            List<Integer> dataIDs = new ArrayList<>(datas.size());
-            Set<Integer> runIds = new HashSet<>();
+            List<Long> dataIDs = new ArrayList<>(datas.size());
+            Set<Long> runIds = new HashSet<>();
             for (ExpData data : datas)
             {
                 dataIDs.add(data.getRowId());
