+++ assay/api-src/org/labkey/api/assay/plate/PlateSetEdge.java
@@ -4,23 +4,23 @@ import java.util.Objects;
 
 public class PlateSetEdge
 {
-    private int _fromPlateSetId;
-    private int _rootPlateSetId;
-    private int _toPlateSetId;
+    private long _fromPlateSetId;
+    private long _rootPlateSetId;
+    private long _toPlateSetId;
 
     // Necessary for database serialization
     public PlateSetEdge()
     {
     }
 
-    public PlateSetEdge(int fromPlateSetId, int toPlateSetId, int rootPlateSetId)
+    public PlateSetEdge(long fromPlateSetId, long toPlateSetId, long rootPlateSetId)
     {
         _fromPlateSetId = fromPlateSetId;
         _toPlateSetId = toPlateSetId;
         _rootPlateSetId = rootPlateSetId;
     }
 
-    public int getFromPlateSetId()
+    public long getFromPlateSetId()
     {
         return _fromPlateSetId;
     }
@@ -30,7 +30,7 @@ public class PlateSetEdge
         _fromPlateSetId = fromPlateSetId;
     }
 
-    public int getRootPlateSetId()
+    public long getRootPlateSetId()
     {
         return _rootPlateSetId;
     }
@@ -40,7 +40,7 @@ public class PlateSetEdge
         _rootPlateSetId = rootPlateSetId;
     }
 
-    public int getToPlateSetId()
+    public long getToPlateSetId()
     {
         return _toPlateSetId;
     }
