+++ assay/src/org/labkey/assay/AssayManager.java
@@ -603,7 +603,7 @@ public class AssayManager implements AssayService
     @Nullable
     public ExpExperiment findBatch(ExpRun run)
     {
-        int protocolId = run.getProtocol().getRowId();
+        long protocolId = run.getProtocol().getRowId();
         for (ExpExperiment potentialBatch : run.getExperiments())
         {
             ExpProtocol batchProtocol = potentialBatch.getBatchProtocol();
@@ -712,7 +712,7 @@ public class AssayManager implements AssayService
         return expRun != null && getProvider(expRun) != null && expRun.getReplacedByRun() == null;
     }
 
-    public void indexAssayBatch(int expRunRowId)
+    public void indexAssayBatch(long expRunRowId)
     {
         ExpRun run = ExperimentService.get().getExpRun(expRunRowId);
         if (run == null)
@@ -768,7 +768,7 @@ public class AssayManager implements AssayService
     }
 
     @Override
-    public void indexAssayRun(int expRunRowId)
+    public void indexAssayRun(long expRunRowId)
     {
         SearchService ss = SearchService.get();
         if (ss == null)
@@ -1031,7 +1031,7 @@ public class AssayManager implements AssayService
     }
 
     @Override
-    public int deleteFlagsForRun(Container container, User user, AssayProvider provider, int runId)
+    public int deleteFlagsForRun(Container container, User user, AssayProvider provider, long runId)
     {
         AssayFlagHandler handler = AssayFlagHandler.getHandler(provider);
         if (handler != null)
@@ -1042,7 +1042,7 @@ public class AssayManager implements AssayService
     }
 
     @Override
-    public <FlagType extends ExpQCFlag> List<FlagType> getFlags(AssayProvider provider, int runId, Class<FlagType> cls)
+    public <FlagType extends ExpQCFlag> List<FlagType> getFlags(AssayProvider provider, long runId, Class<FlagType> cls)
     {
         AssayFlagHandler handler = AssayFlagHandler.getHandler(provider);
         if (handler != null)
